Split native click-probe validator
