+++ docs/DECISIONS.md
@@ -2,6 +2,7 @@
 
 ## 2026-07-01
 
+- Native click-probe validation should keep the public script path stable while splitting implementation by release artifact boundary. `scripts/native-click-probe-contracts.py` is now only the executable facade used by smoke scripts; `scripts/native_click_probe_contracts/` owns constants, JSON/value helpers, click-probe normalization, packaged-window comparison/readiness manifests, live Accessibility frame validation, and CLI wiring. Parity tests read the full validator package instead of the facade, so release contract checks still fail if selector precedence, sample coordinates, policy drift, window command coverage, or live frame evidence regresses.
 - Native text-entry click targets must be addressable by Accessibility ID, not only by internal focus state. The source hit-target gate now rejects raw `TextField`, `SecureField`, and `TextEditor` controls unless the owning control scope declares `.accessibilityIdentifier(...)` beside the `quillCodeTextEntryTarget()`. Reusable dialog text fields synthesize deterministic `quillcode-...-field` identifiers, while specialized settings, transcript-find, browser, terminal, and review/PR fields use explicit IDs. This closes the "field looks right but cannot be clicked or typed into by smoke automation" gap without weakening the existing 44 pt/focus-target contracts.
 - Instruction Review dismissals and resolved-by-edit audit records are project-scoped diagnostic records, not only pane UI state. `ProjectRef` owns normalized `ProjectInstructionDiagnosticResolution` records with `dismissed` and `resolved` dispositions, metadata refresh preserves them, and Activity combines durable dismissals with transient session dismissals before rendering diagnostics. A diagnostic that disappears after a context refresh is recorded as `resolved`; that is audit history only and never suppresses a future reintroduced diagnostic with the same ID. Only `dismissed` hides currently active diagnostics. The dismiss command only accepts currently known instruction diagnostics, so stale or fabricated command IDs cannot hide nonexistent rule issues. Resolve remains an explicit draft/edit workflow because changing `AGENTS.md` or `.quillcode/rules.md` should leave a normal file-edit audit trail.
