+++ scripts/native_click_probe_contracts/__init__.py
@@ -1,0 +1,1 @@
+"""Native click-probe contract validation helpers."""
